user_system/requirements.txt
# Django Core
Django==5.1.4
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Django Extensions & Utilities
django-extensions==3.2.3

# Authentication & Security
django-allauth==0.63.6
django-otp==1.5.4
django-axes==6.4.0
django-ratelimit==4.1.0
cryptography==43.0.3

# API & REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.3
django-cors-headers==4.4.0

# Task Queue & Cache
celery==5.4.0
redis==5.2.0

# Image Processing
Pillow==10.4.0

# QR Code Generation
qrcode[pil]==8.0

# Production Server
gunicorn==23.0.0

# Timezone support
pytz==2024.2

# Development & Testing
django-debug-toolbar==4.4.6